extract eventtype logic
